#224 #233: test files named for the contract they pin; migration 0008 docstring tells the truth
Renames, every assertion kept:
- test_pass4_engine_pins.py    -> test_engine_failure_paths.py
- test_pass4_testing.py        -> test_testing_helper_guards.py
- test_pass4_checks.py         -> test_settings_and_shadowing_checks.py
- test_issue_fixes_engine.py   -> test_lock_release_and_restore_pins.py
- test_issue_fixes_testing.py  -> test_testing_contract_pins.py
- stability/test_periodic_tasks.py -> test_safety_net_contracts.py,
  retargeted at the claim filter and the two remaining safety nets

test_issue_fixes_0_12.py split along its own section markers into
test_write_failure_accounting.py (the failing-write accounting pins)
and test_definition_validation_pins.py (definition/binding/settings
refusals and the public surface).

Migration 0008 is named after columns 0009 removed; renaming an applied
migration would ghost it in every consumer's django_migrations, so the
docstring now says which column is live and the rename waits for the
1.0 squash.
