+++ tests/test_lock_release_and_restore_pins.py
